screen recorder/requirements.txt
flask==2.0.1
selenium==4.1.0
webdriver-manager==3.5.2
beautifulsoup4==4.10.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
scikit-image==0.18.3
opencv-python==4.5.3.56
pillow==8.3.2